Prevent non-affiliated records from loading
KOHA-1492: Sometimes we get records from Libris where there is no
affiliation for our library. The cause for this is unknown.

This patch raises a NonLibraryAffiliatedRecord error which is then
caught and ignored to prevent such records from being loaded into
the library system.

Affiliation is defined as records containing one or more 852 fields.

This raises the error when 003 is set to LIBRIS, and no 852 field
is present.

Loading records from other sources than Libris should therefor be
allowed.