KOHA-1510 Insert dash in ISSN for field 035$a
Before LibrisXL 035$9 had old information identifying the record.
This could be multiple different types of identifiers, where ISSN
was one of them. This identifier was stored without punctuation.

Now this field has been moved to 035$a instead, and adjustlibris
needs to handle both cases. The logic is the same except for the
subfield. If the value is 8 characters, put a dash in the middle,
otherwise keep it as it is.

Ex.
12345678 becomes 1234-5678